Treat matlab less janky.
Before, only attempt loading matlab is on mac and matlab installed right
place.

Now, load matlab if the matlab command can be found, make sure if can
be found on Mac by adding Mac bin path only if on mac and it exists.

The on Mac? predicae may still be too much.